ats_web/backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.31.1
python-multipart>=0.0.9
PyPDF2==3.0.1
openai>=1.66.0
pydantic>=2.11.0,<3.0.0
python-dotenv>=1.0.1
anyio>=4.7.0
h11>=0.16.0
tenacity>=8.1.0,<9.0.0,!=8.4.0
packaging>=23.2,<25
openpyxl==3.1.2

# Feedback & Vector Database
chromadb>=0.4.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
numpy>=1.24.0
